pyproject.toml
[tool.poetry]
name = "fashion-mnist"
version = "0.1.0"
description = "Fashion MNIST with Pytorch CNN"
authors = ["Kravchenko01 <roma.kravchenco.2001@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.1.1"
torchvision = "^0.16.1"
dvc = "^3.38.1"
hydra-core = "^1.3.2"
lightning = "^2.1.3"
torchmetrics = "^1.3.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.6.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.black]
line-length = 90
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
max-line-length = 90
ignore = ['E203', 'E501', 'W503', 'B950']
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']
per-file-ignores = [
    # for easier imports to __init__ without __all__
    '**/__init__.py: F401',
    # file to define custom types
    'types.py: F401'
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
